Add stylesheet directly to plugin view context.
Before, Timber was only sending the plugin directory in the context to the view to render the zoneboard. The problem arises if the plugin folder name changes. This happened to me when I downloaded the plugin from GitHub and it had the name "zoneboard-master." The view is looking in the "zoneboard" folder within the plugins directory, which obviously didn't exist in my case. Now I have added the URL to the file to the context (with a more semantic name) so that it doesn't matter what the plugin folder is called, as long as the internal structure remains unchanged.